fix(vercel): move i18n out of internal/ so @vercel/go builds the handler
@vercel/go recompiles the entrypoint under a synthetic package path
("handler/vercel"), from which a github.com/.../internal/* package is
off-limits. So vercel/handler.go importing internal/i18n directly broke
the production build ("use of internal package ... not allowed") — this
is what failed the deploy of the previous merge.

Move i18n from internal/i18n to a top-level i18n package (it's a reusable
leaf lib, not project-private) and rewrite every import. The handler now
imports only non-internal packages; internal db/handlers stay put (only
reached transitively via app/, which Go allows).

No behavior change: build/vet/test green, render unchanged EN/ES/FR.
