+++ app/components/aslanImage.templ
@@ -3,7 +3,7 @@ package components
 import (
 	"github.com/AslanSN/CurriculumVitae/helpers"
 	"github.com/AslanSN/CurriculumVitae/internal/db/constants"
-	"github.com/AslanSN/CurriculumVitae/internal/i18n"
+	"github.com/AslanSN/CurriculumVitae/i18n"
 )
 
 templ AslanImage() {
