+++ app/components/contactModalFooter.templ
@@ -2,7 +2,7 @@ package components
 
 import (
 	"github.com/AslanSN/CurriculumVitae/helpers"
-	"github.com/AslanSN/CurriculumVitae/internal/i18n"
+	"github.com/AslanSN/CurriculumVitae/i18n"
 	"net/url"
 	"strings"
 )
