+++ app/components/experienceCard_templ.go
@@ -9,8 +9,8 @@ import "github.com/a-h/templ"
 import templruntime "github.com/a-h/templ/runtime"
 
 import (
+	"github.com/AslanSN/CurriculumVitae/i18n"
 	"github.com/AslanSN/CurriculumVitae/internal/db/constants"
-	"github.com/AslanSN/CurriculumVitae/internal/i18n"
 )
 
 func ExperienceCard(workplace constants.ExperienceStruct) templ.Component {
