+++ app/components/hero.templ
@@ -1,6 +1,6 @@
 package components
 
-import "github.com/AslanSN/CurriculumVitae/internal/i18n"
+import "github.com/AslanSN/CurriculumVitae/i18n"
 
 templ Hero() {
 	<section id="hero" class="flex scroll-mt-24 flex-col items-start gap-8 pt-6 md:flex-row md:items-center md:justify-between md:gap-12">
