+++ app/components/navbar.templ
@@ -3,7 +3,7 @@ package components
 import (
 	icons "github.com/AslanSN/CurriculumVitae/app/components/iconComponents"
 	"github.com/AslanSN/CurriculumVitae/helpers"
-	"github.com/AslanSN/CurriculumVitae/internal/i18n"
+	"github.com/AslanSN/CurriculumVitae/i18n"
 )
 
 // navVisibility hides the horizontal navbar until the viewport is wide enough
