+++ app/components/sidebar.templ
@@ -3,7 +3,7 @@ package components
 import (
 	icons "github.com/AslanSN/CurriculumVitae/app/components/iconComponents"
 	"github.com/AslanSN/CurriculumVitae/helpers"
-	"github.com/AslanSN/CurriculumVitae/internal/i18n"
+	"github.com/AslanSN/CurriculumVitae/i18n"
 )
 
 // Sidebar is the mobile navigation drawer (CSS-only open/close via a peer
