+++ app/components/skills.templ
@@ -2,7 +2,7 @@ package components
 
 import (
 	"github.com/AslanSN/CurriculumVitae/internal/db/constants"
-	"github.com/AslanSN/CurriculumVitae/internal/i18n"
+	"github.com/AslanSN/CurriculumVitae/i18n"
 )
 
 templ Skills() {
