+++ app/views/home.templ
@@ -4,7 +4,7 @@ import (
 	"github.com/AslanSN/CurriculumVitae/app/components"
 	"github.com/AslanSN/CurriculumVitae/app/layouts"
 	"github.com/AslanSN/CurriculumVitae/internal/db/constants"
-	"github.com/AslanSN/CurriculumVitae/internal/i18n"
+	"github.com/AslanSN/CurriculumVitae/i18n"
 )
 
 templ home() {
