+++ app/views/index.templ
@@ -1,6 +1,6 @@
 package views
 
-import "github.com/AslanSN/CurriculumVitae/internal/i18n"
+import "github.com/AslanSN/CurriculumVitae/i18n"
 
 templ Index() {
 	<!DOCTYPE html>
