+++ internal/db/constants/challenges.go
@@ -1,7 +1,7 @@
 package constants
 
 import (
-	"github.com/AslanSN/CurriculumVitae/internal/i18n"
+	"github.com/AslanSN/CurriculumVitae/i18n"
 	"github.com/a-h/templ"
 )
