+++ internal/db/constants/experience.go
@@ -2,7 +2,7 @@ package constants
 
 import (
 	"github.com/AslanSN/CurriculumVitae/helpers"
-	"github.com/AslanSN/CurriculumVitae/internal/i18n"
+	"github.com/AslanSN/CurriculumVitae/i18n"
 	"github.com/a-h/templ"
 )
